reorganizando estrutura de banco de dados no backend e migrando automaticamente com flyway
AI-tool: ChatGPT
AI-prompt: "Como migrar o banco de dados mariadb automaticamente usando ktor"
