+++ backend/build.gradle.kts
@@ -33,5 +33,6 @@ dependencies {
     implementation("org.jetbrains.exposed:exposed-core:0.41.1")
     implementation("org.jetbrains.exposed:exposed-dao:0.41.1")
     implementation("org.jetbrains.exposed:exposed-jdbc:0.41.1")
-    implementation("org.mariadb.jdbc:mariadb-java-client:3.2.0")
+    implementation("org.mariadb.jdbc:mariadb-java-client:3.0.7")
+    implementation("org.flywaydb:flyway-core:8.0.0")
 }
