+++ backend/src/main/kotlin/Application.kt
@@ -2,6 +2,8 @@ package com.bibliophile
 
 import com.bibliophile.config.*
 import io.ktor.server.application.*
+
+import com.bibliophile.db.DatabaseFactory
 import com.bibliophile.repositories.BooklistRepository
 
 fun main(args: Array<String>) {
@@ -14,7 +16,8 @@ fun Application.module() {
     configureSerialization(repository)
 
     configureMonitoring()
-    configureDatabases()
     configureRouting()
     configureHTTP()
+
+    DatabaseFactory.init()
 }
