+++ backend/src/main/kotlin/db/DaoMapper.kt
@@ -1,0 +1,11 @@
+package com.bibliophile.db
+
+import com.bibliophile.models.Booklist
+import com.bibliophile.db.entities.BooklistDAO
+
+fun daoToModel(dao: BooklistDAO) = Booklist(
+    dao.id.value,
+    dao.userId,
+    dao.name,
+    dao.description ?: "",
+)
