+++ backend/src/main/kotlin/db/Transactions.kt
@@ -1,0 +1,8 @@
+package com.bibliophile.db
+
+import kotlinx.coroutines.Dispatchers
+import org.jetbrains.exposed.sql.Transaction
+import org.jetbrains.exposed.sql.transactions.experimental.newSuspendedTransaction
+
+suspend fun <T> suspendTransaction(block: Transaction.() -> T): T =
+    newSuspendedTransaction(Dispatchers.IO, statement = block)
